Refactor mask creation: consolidate duplicate code and modernize type hints
