+++ namedivider/feature/extractor.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 import namedivider.feature.functional as F
 from namedivider.feature.family_name import FamilyNameRepository
 from namedivider.feature.kanji import KanjiStatisticsRepository
-from namedivider.feature.mask_cache import MaskCache
+from namedivider.feature.functional import MaskCache
 
 
 @dataclass(frozen=True)
